explicit error for long-vector inputs
git-svn-id: https://svn.r-project.org/R/trunk@65956 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/main/util.c
@@ -607,7 +607,7 @@ SEXP attribute_hidden do_merge(SEXP call, SEXP op, SEXP args, SEXP rho)
 	for(; j < ny; j++) if(INTEGER(yi)[j] >= tmp) break;
 	for(nny = j; nny < ny; nny++) if(INTEGER(yi)[nny] != tmp) break;
 	/* printf("i %d nnx %d j %d nny %d\n", i, nnx, j, nny); */
-	dnans += (nnx-i)*(nny-j);
+	dnans += ((double)(nnx-i))*(nny-j);
     }
     if (dnans > R_XLEN_T_MAX)
 	error(_("number of rows in the result exceeds maximum vector length"));
